% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{clustdiffgenes}
\alias{clustdiffgenes}
\alias{clustdiffgenes,SCseq-method}
\title{clustdiffgenes}
\usage{
clustdiffgenes(object, pvalue = 0.01)

\S4method{clustdiffgenes}{SCseq}(object, pvalue = 0.01)
}
\description{
clustdiffgenes
}
\section{Methods (by class)}{
\itemize{
\item \code{SCseq}: 
}}

