% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{clustexp}
\alias{clustexp}
\alias{clustexp,SCseq-method}
\title{clustexp}
\usage{
clustexp(object, clustnr = 20, bootnr = 50, metric = "pearson",
  do.gap = TRUE, SE.method = "Tibs2001SEmax", SE.factor = 0.25,
  B.gap = 50, cln = 0, rseed = 17000)

\S4method{clustexp}{SCseq}(object, clustnr = 20, bootnr = 50,
  metric = "pearson", do.gap = TRUE, SE.method = "Tibs2001SEmax",
  SE.factor = 0.25, B.gap = 50, cln = 0, rseed = 17000)
}
\description{
clustexp
}
\section{Methods (by class)}{
\itemize{
\item \code{SCseq}: 
}}

