% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{filterdata}
\alias{filterdata}
\alias{filterdata,SCseq-method}
\title{filterdata}
\usage{
filterdata(object, mintotal = 3000, minexpr = 5, minnumber = 1,
  maxexpr = 500, downsample = FALSE, dsn = 1, rseed = 17000)

\S4method{filterdata}{SCseq}(object, mintotal = 3000, minexpr = 5,
  minnumber = 1, maxexpr = 500, downsample = FALSE, dsn = 1,
  rseed = 17000)
}
\description{
filterdata
}
\section{Methods (by class)}{
\itemize{
\item \code{SCseq}: 
}}

