% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class.r
\docType{class}
\name{SCseq-class}
\alias{SCseq}
\alias{SCseq-class}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{expdata}}{}

\item{\code{ndata}}{}

\item{\code{fdata}}{}

\item{\code{distances}}{}

\item{\code{tsne}}{}

\item{\code{kmeans}}{}

\item{\code{background}}{}

\item{\code{out}}{}

\item{\code{cpart}}{}

\item{\code{fcol}}{}

\item{\code{filterpar}}{}

\item{\code{clusterpar}}{}

\item{\code{outlierpar}}{}
}}

